+++ src/app/app.routing.ts
@@ -31,7 +31,7 @@ import { UserListComponent } from "./core/admin/user-list/user-list.component";
 import { EntityDetailsComponent } from "./core/entity-components/entity-details/entity-details.component";
 import { ConflictResolutionListComponent } from "./conflict-resolution/conflict-resolution-list/conflict-resolution-list.component";
 import { ActivityListComponent } from "./child-dev-project/attendance/activity-list/activity-list.component";
-import { ReportingComponent } from "./child-dev-project/reporting/reporting/reporting.component";
+import { ReportingComponent } from "./features/reporting/reporting/reporting.component";
 
 export const COMPONENT_MAP = {
   Dashboard: DashboardComponent,
