+++ src/app/features/reporting/reporting/report-row/report-row.component.scss
