+++ src/app/features/reporting/reporting/reporting.component.scss
